+++ src/debugger/evalstackmachine.h
@@ -82,6 +82,7 @@ struct EvalData
     ToRelease<ICorDebugClass> trVoidClass;
     std::unordered_map<CorElementType, ToRelease<ICorDebugClass>> trElementToValueClassMap;
     FrameLevel frameLevel;
+    FormatSpecifier specifier{FormatSpecifier::None};
 };
 
 class EvalStackMachine
@@ -111,8 +112,9 @@ class EvalStackMachine
     }
 
     // Evaluate expression. Optional, return `editable` state and in case result is property - setter related information.
-    HRESULT EvaluateExpression(ICorDebugThread *pThread, FrameLevel frameLevel, const std::string &expression, ICorDebugValue **ppResultValue,
-                               std::string &output, bool *editable = nullptr, std::unique_ptr<Evaluator::SetterData> *resultSetterData = nullptr);
+    HRESULT EvaluateExpression(ICorDebugThread *pThread, FrameLevel frameLevel, const std::string &expression,
+                               FormatSpecifier specifier, ICorDebugValue **ppResultValue, std::string &output,
+                               bool *editable = nullptr, std::unique_ptr<Evaluator::SetterData> *resultSetterData = nullptr);
 
     // Set value in pValue by expression with implicitly cast expression result to pValue type, if need.
     HRESULT SetValueByExpression(ICorDebugThread *pThread, FrameLevel frameLevel, ICorDebugValue *pValue,